Moving this schedule out of the adherence package as it no longer includes adherence information
