+++ src/main/java/org/sagebionetworks/bridge/services/Schedule2Service.java
@@ -48,8 +48,8 @@ import org.sagebionetworks.bridge.models.schedules2.Schedule2;
 import org.sagebionetworks.bridge.models.schedules2.Session;
 import org.sagebionetworks.bridge.models.schedules2.TimeWindow;
 import org.sagebionetworks.bridge.models.schedules2.adherence.AdherenceState;
-import org.sagebionetworks.bridge.models.schedules2.adherence.participantschedule.ParticipantSchedule;
-import org.sagebionetworks.bridge.models.schedules2.adherence.participantschedule.ParticipantScheduleGenerator;
+import org.sagebionetworks.bridge.models.schedules2.participantschedules.ParticipantSchedule;
+import org.sagebionetworks.bridge.models.schedules2.participantschedules.ParticipantScheduleGenerator;
 import org.sagebionetworks.bridge.models.schedules2.timelines.Scheduler;
 import org.sagebionetworks.bridge.models.schedules2.timelines.Timeline;
 import org.sagebionetworks.bridge.models.schedules2.timelines.TimelineMetadata;
