+++ src/test/java/org/sagebionetworks/bridge/models/schedules2/adherence/participantschedule/ParticipantScheduleGeneratorTest.java
@@ -3,7 +3,7 @@ package org.sagebionetworks.bridge.models.schedules2.adherence.participantschedu
 import static org.sagebionetworks.bridge.TestConstants.SESSION_GUID_1;
 import static org.sagebionetworks.bridge.TestConstants.TIMESTAMP;
 import static org.sagebionetworks.bridge.TestConstants.TIMEZONE_MSK;
-import static org.sagebionetworks.bridge.models.schedules2.adherence.participantschedule.ParticipantScheduleGenerator.INSTANCE;
+import static org.sagebionetworks.bridge.models.schedules2.participantschedules.ParticipantScheduleGenerator.INSTANCE;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNull;
 import static org.testng.Assert.assertTrue;
@@ -14,6 +14,7 @@ import org.sagebionetworks.bridge.models.activities.StudyActivityEvent;
 import org.sagebionetworks.bridge.models.schedules2.Schedule2;
 import org.sagebionetworks.bridge.models.schedules2.Schedule2Test;
 import org.sagebionetworks.bridge.models.schedules2.adherence.AdherenceState;
+import org.sagebionetworks.bridge.models.schedules2.participantschedules.ParticipantSchedule;
 import org.sagebionetworks.bridge.models.schedules2.timelines.ScheduledSession;
 import org.sagebionetworks.bridge.models.schedules2.timelines.Scheduler;
 import org.sagebionetworks.bridge.models.schedules2.timelines.Timeline;
